# api/openapi.yaml
openapi: 3.0.3
info:
  title: bluebell-new API
  version: 1.0.0
  description: Bluebell Gin HTTP API contract maintained independently from handler comments.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /refresh_token:
    post:
      tags: [auth]
      summary: 刷新访问令牌
      description: 使用 refresh_token 轮换并返回新的 access_token 与 refresh_token。该接口不需要 Bearer Token。
      operationId: refreshToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
            examples:
              valid:
                summary: 有效刷新参数
                value:
                  refresh_token: eyJhbGciOi...
      responses:
        '200':
          description: 刷新结果。业务错误也以 HTTP 200 返回，并通过 code 区分。
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/APIResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/RefreshTokenResponse'
              examples:
                success:
                  summary: 刷新成功
                  value:
                    code: 0
                    message: success
                    data:
                      access_token: eyJhbGciOi...
                      refresh_token: eyJhbGciOi...
                request_error:
                  summary: 参数校验失败
                  value:
                    code: -400
                    message: 请求错误
                    data: null
                invalid_refresh_token:
                  summary: refresh_token 无效
                  value:
                    code: 10006
                    message: 刷新token无效
                    data: null
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    APIResponse:
      type: object
      required: [code, message, data]
      properties:
        code:
          type: integer
          format: int32
          example: 0
        message:
          type: string
          example: success
        data:
          nullable: true
...
    RefreshTokenRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
          minLength: 1
          description: 登录或上次刷新得到的 refresh_token。
    RefreshTokenResponse:
      type: object
      required: [access_token, refresh_token]
      properties:
        access_token:
          type: string
          description: 新的 JWT access token。
        refresh_token:
          type: string
          description: 新的 JWT refresh token；旧 refresh_token 会失效。
...